+++ spec/integration/keys_spec.rb
@@ -36,4 +36,39 @@ describe "Definition.Keys" do
                     },
                     receive_newsletter: false
                   }
+
+  context "with unexpected key in input hash" do
+    subject(:definition) do
+      Definition.Keys do
+        required :first_name, Definition.Type(String)
+      end
+    end
+
+    let(:value) do
+      {
+        first_name:     "Jon",
+        unexpected_key: "foobar"
+      }
+    end
+
+    it_behaves_like "it does not conform"
+  end
+
+  context "with unexpected key in input hash and option to ignore unexpected key set to true" do
+    subject(:definition) do
+      Definition.Keys do
+        option :ignore_extra_keys
+        required :first_name, Definition.Type(String)
+      end
+    end
+
+    it_behaves_like "it conforms via coersion",
+                    input:  {
+                      first_name:     "Jon",
+                      unexpected_key: "foobar"
+                    },
+                    output: {
+                      first_name: "Jon"
+                    }
+  end
 end