+++ Changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-03-28
+### Added
+- CoercibleValueObject Definition for better nesting of ValueObjects
+- Nilable Definition as shortcut for nil OR some other definition
+- Option for Keys Definition to ignore unexpected keys
+### Fixed
+- Error hash was missing some errors in a few cases
+- `ConformResult.error_hash` was empty for Definitions without any `Keys` Definition
+### Changed
+- And Definition stops processing after the first failure
+
 ## [0.4.0] - 2019-03-22
 ### Added
 - Added support for default values to Keys Definition